requirements.txt
streamlit
google-generativeai
python-dotenv
pdfplumber
matplotlib
pandas
scikit-learn